pyproject.toml
[project]
name = "langchain-academy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncer>=0.0.17",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.13",
    "langchain-mcp-adapters>=0.2.1",
    "langchain-openai>=1.1.9",
    "langchain-openrouter>=0.0.2",
    "langchain-tavily>=0.2.17",
    "langgraph>=1.0.8",
    "langgraph-checkpoint-sqlite>=3.0.3",
    "langgraph-cli[inmem]>=0.4.12",
    "langgraph-prebuilt>=1.0.7",
    "langgraph-sdk>=0.3.6",
    "langsmith>=0.7.3",
    "notebook>=7.5.3",
    "rich[jupyter]>=14.3.2",
    "trustcall>=0.0.39",
    "wikipedia>=1.4.0",
]

[dependency-groups]
dev = []
